+++ allauth/mfa/models.py
@@ -1,5 +1,6 @@
 from django.conf import settings
 from django.db import models
+from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
 
@@ -20,6 +21,8 @@ class Authenticator(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     type = models.CharField(max_length=20, choices=Type.choices)
     data = models.JSONField()
+    created_at = models.DateTimeField(default=timezone.now)
+    last_used_at = models.DateTimeField(null=True)
 
     class Meta:
         unique_together = (("user", "type"),)
@@ -34,3 +37,7 @@ class Authenticator(models.Model):
         }[
             self.type
         ](self)
+
+    def record_usage(self):
+        self.last_used_at = timezone.now()
+        self.save(update_fields=["last_used_at"])